# swagger.yml
openapi: 3.0.2
info:
  title: API Ercole 2.0
  version: 1.0.0
  license:
    name: GNU General Public License v3.0 or later
    url: 'https://www.gnu.org/licenses/gpl-3.0-standalone.html'
  description: API used to talk to ercole services
  contact:
    name: Issues support
    url: 'https://github.com/ercole-io/ercole'
servers:
  - description: Api service on dev.ercole.io
    url: 'https://dev.ercole.io/api'
  - description: Data service on dev.ercole.io
    url: 'https://dev.ercole.io/data'
  - description: Alert service on dev.ercole.io
    url: 'https://dev.ercole.io/alert'
  - description: Repo service on dev.ercole.io
    url: 'https://dev.ercole.io/repo'
  - description: localhost address
    url: 'http://localhost:11111'
...
components:
  securitySchemes:
    basicAuthDataService:
      type: http
      scheme: basic
      description: access to data-service APIs
    basicAuthAlertService:
      type: http
      scheme: basic
      description: access to alert-service APIs
    basicAuthApiService:
      type: http
      scheme: basic
      description: access to api-service APIs
    tokenAuthApiService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access using the access token
  responses:
    error:
      description: Contains some informations about a error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required:
              - Error
              - ErrorDescription
              - SourceFilename
              - LineNumber
            properties:
              Error:
                type: string
              ErrorDescription:
                type: string
              SourceFilename:
                type: string
              LineNumber:
                type: integer
          examples:
            Test1:
              value:
                Error: Invalid  status
                ErrorDescription: invalid status
                SourceFilename: /home/travis/go/src/github.com/ercole-io/ercole/api-service/controller/alerts_api.go
                LineNumber: 65
...
  parameters:
...
    location:
      schema:
        type: string
        example: Italy
      in: query
      name: location
      description: Filter by location
      allowEmptyValue: true
    environment:
      schema:
        type: string
        example: TST
      in: query
      name: environment
      description: Filter by environment
      allowEmptyValue: true
    older-than:
      schema:
        type: string
        format: date-time
        example: '2020-05-05T08:23:40+00:00'
      in: query
      name: older-than
      description: Filter until the date
      allowEmptyValue: true
...
paths:
...
  /hosts/count:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetHostsCountStats
      summary: Get the number of the hosts
      description: Get the number of the hosts
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: integer
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  /hosts/environments/frequency:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetEnvironmentStats
      summary: Get the stats about the environments
      description: Get the stats about the environments
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    Environment:
                      type: string
                    Count:
                      type: integer
                  required:
                    - Environment
                    - Count
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  /hosts/types:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetTypeStats
      summary: Get the stats about the types
      description: Get the stats about the types
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    HardwareAbstractionTechnology:
                      type: string
                    Count:
                      type: integer
                  required:
                    - HardwareAbstractionTechnology
                    - Count
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
  /hosts/operating-systems:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetOperatingSystemStats
      summary: Get the stats about the operating systems 
      description: Get the stats about the operating systems
      parameters:
        - $ref: '#/components/parameters/location'
        - $ref: '#/components/parameters/environment'
        - $ref: '#/components/parameters/older-than'
      responses:
        '200':
          description: statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    OperatingSystem:
                      type: string
                    Count:
                      type: integer
                  required:
                    - OperatingSystem
                    - Count
        '401':
          $ref: '#/components/responses/error'
        '422':
          $ref: '#/components/responses/error'
        '500':
          $ref: '#/components/responses/error'
...
security:
  - basicAuthApiService: []
  - tokenAuthApiService: []